requirements.txt
kafka-python
faker
Flask
flask_apscheduler